+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `isValidRepositoryName` function
 - `submitPath` function
+- `Enter` key shortcut
 ### Changed
 - Repository name bug fixed
 - `parseGithubPath` function modified